refactor: rename pspline module
