+++ moderndid/didcont/npiv/cck_ucb.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from ..utils import _quantile_basis, avoid_zero_division
 from .estimators import _ginv, npiv_est
-from .pspline import prodspline
+from .prodspline import prodspline
 from .results import NPIVResult
